// src/type.ts
export type StatusValue = 'init' | 'play' | 'bomb' | 'clear'

export type SizeValue = 'small' | 'middle' | 'large'

export type DensityValue = 'low' | 'middle' | 'high'

export type DotValue =
  | ''
  | 'flag'
  | '0'
  | '1'
  | '2'
  | '3'
  | '4'
  | '5'
  | '6'
  | '7'
  | '8'

export type SweptValue = 'safe' | 'bomb' | 'stay' | 'clear'